Created a few more repositories and made some changes to the main repository class.
Created repositories for the Author, AuthorBook and Book entities. Added interfaces for all repositories. Added validation for the properties that should be included for eager loading in the GetAsync() method. Made some changes to the abstract Repository class, now there is no need to specify the type of the entity`s primary key, this led to some difficulties with creating a mapping table and a corresponding repository. Every repository will implement their own GetByIdAsync() method, it is no longer present in the main Repository class. Updated both the interface IUnitOfWork and class UnitOfWork.