Removed the 'Discriminator' column from the database and made the User's Id to be a GUID, not a string. Removed some of the default identity pages, which were previously generated.
Due to the previous implementation of the ApplicationUser not being quite correct ( it was partial ) - in some places the default IdentityUser was used, in others the ApplicationUser and this led to a column 'Discriminator' being created in the database ( due to polymorphism ), which is not good to have. Since a lot of changes had to be made to the application in order to remove the 'Discriminator' column, I decided to change the User's id as well.